BF transpilation, with issues.

+++ src/stack/mod.rs
@@ -8,5 +8,5 @@ pub use inst::*;
 
 pub use super::*;
 
-pub type Word = u32;
+pub type Word = u8;
 pub const WORD_SIZE: usize = 4;
